Rename science directories to science_and_technology
