+++ lib/logic/tools/science_and_technology/irrational_numbers/pi.dart
@@ -1,4 +1,4 @@
-import 'package:gc_wizard/logic/tools/science/irrational_numbers/irrational_numbers.dart';
+import 'package:gc_wizard/logic/tools/science_and_technology/irrational_numbers/irrational_numbers.dart';
 
 final PI = IrrationalNumber(
   symbol : String.fromCharCode(960),
