+++ lib/logic/tools/science_and_technology/numeral_bases.dart
