+++ lib/widgets/registry.dart
@@ -76,16 +76,16 @@ import 'package:gc_wizard/widgets/tools/encodings/roman_numbers.dart';
 import 'package:gc_wizard/widgets/tools/encodings/scrabble.dart';
 import 'package:gc_wizard/widgets/tools/encodings/z22.dart';
 import 'package:gc_wizard/widgets/tools/formula_solver/formula_solver.dart';
-import 'package:gc_wizard/widgets/tools/science/irrational_numbers/e.dart';
-import 'package:gc_wizard/widgets/tools/science/irrational_numbers/phi.dart';
-import 'package:gc_wizard/widgets/tools/science/irrational_numbers/pi.dart';
-import 'package:gc_wizard/widgets/tools/science/numeralbases.dart';
-import 'package:gc_wizard/widgets/tools/science/periodic_table.dart';
-import 'package:gc_wizard/widgets/tools/science/primes/primes_integerfactorization.dart';
-import 'package:gc_wizard/widgets/tools/science/primes/primes_isprime.dart';
-import 'package:gc_wizard/widgets/tools/science/primes/primes_nearestprime.dart';
-import 'package:gc_wizard/widgets/tools/science/primes/primes_nthprime.dart';
-import 'package:gc_wizard/widgets/tools/science/primes/primes_primeindex.dart';
+import 'package:gc_wizard/widgets/tools/science_and_technology/irrational_numbers/e.dart';
+import 'package:gc_wizard/widgets/tools/science_and_technology/irrational_numbers/phi.dart';
+import 'package:gc_wizard/widgets/tools/science_and_technology/irrational_numbers/pi.dart';
+import 'package:gc_wizard/widgets/tools/science_and_technology/numeralbases.dart';
+import 'package:gc_wizard/widgets/tools/science_and_technology/periodic_table.dart';
+import 'package:gc_wizard/widgets/tools/science_and_technology/primes/primes_integerfactorization.dart';
+import 'package:gc_wizard/widgets/tools/science_and_technology/primes/primes_isprime.dart';
+import 'package:gc_wizard/widgets/tools/science_and_technology/primes/primes_nearestprime.dart';
+import 'package:gc_wizard/widgets/tools/science_and_technology/primes/primes_nthprime.dart';
+import 'package:gc_wizard/widgets/tools/science_and_technology/primes/primes_primeindex.dart';
 
 class Registry {
   static List<GCWToolWidget> toolList;
