--- lib/widgets/tools/science/irrational_numbers/e.dart
@@ -1,16 +1,0 @@
-import 'package:gc_wizard/logic/tools/science/irrational_numbers/e.dart';
-import 'package:gc_wizard/widgets/tools/science/irrational_numbers/irrationalnumbers_decimalrange.dart';
-import 'package:gc_wizard/widgets/tools/science/irrational_numbers/irrationalnumbers_nthdecimal.dart';
-import 'package:gc_wizard/widgets/tools/science/irrational_numbers/irrationalnumbers_search.dart';
-
-class ENthDecimal extends IrrationalNumbersNthDecimal {
-  ENthDecimal() : super(irrationalNumber: E);
-}
-
-class EDecimalRange extends IrrationalNumbersDecimalRange {
-  EDecimalRange() : super(irrationalNumber: E);
-}
-
-class ESearch extends IrrationalNumbersSearch {
-  ESearch() : super(irrationalNumber: E);
-}
