+++ lib/widgets/tools/science_and_technology/irrational_numbers/e.dart
@@ -1,0 +1,16 @@
+import 'package:gc_wizard/logic/tools/science_and_technology/irrational_numbers/e.dart';
+import 'package:gc_wizard/widgets/tools/science_and_technology/irrational_numbers/irrationalnumbers_decimalrange.dart';
+import 'package:gc_wizard/widgets/tools/science_and_technology/irrational_numbers/irrationalnumbers_nthdecimal.dart';
+import 'package:gc_wizard/widgets/tools/science_and_technology/irrational_numbers/irrationalnumbers_search.dart';
+
+class ENthDecimal extends IrrationalNumbersNthDecimal {
+  ENthDecimal() : super(irrationalNumber: E);
+}
+
+class EDecimalRange extends IrrationalNumbersDecimalRange {
+  EDecimalRange() : super(irrationalNumber: E);
+}
+
+class ESearch extends IrrationalNumbersSearch {
+  ESearch() : super(irrationalNumber: E);
+}
