+++ lib/widgets/tools/science_and_technology/irrational_numbers/irrationalnumbers_search.dart
@@ -1,16 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:gc_wizard/i18n/app_localizations.dart';
-import 'package:gc_wizard/logic/tools/crypto/rotator.dart';
-import 'package:gc_wizard/logic/tools/science/irrational_numbers/irrational_numbers.dart';
-import 'package:gc_wizard/logic/tools/science/primes/primes.dart';
+import 'package:gc_wizard/logic/tools/science_and_technology/irrational_numbers/irrational_numbers.dart';
 import 'package:gc_wizard/utils/constants.dart';
-import 'package:gc_wizard/widgets/common/base/gcw_textfield.dart';
-import 'package:gc_wizard/widgets/common/gcw_integer_spinner.dart';
 import 'package:gc_wizard/widgets/common/gcw_default_output.dart';
 import 'package:gc_wizard/widgets/common/gcw_integer_textfield.dart';
-import 'package:gc_wizard/widgets/common/gcw_text_divider.dart';
-import 'package:gc_wizard/widgets/utils/common_widget_utils.dart';
-import 'package:sqflite/sqlite_api.dart';
 
 class IrrationalNumbersSearch extends StatefulWidget {
   final IrrationalNumber irrationalNumber;
