+++ lib/widgets/tools/science_and_technology/numeralbases.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:gc_wizard/i18n/app_localizations.dart';
-import 'package:gc_wizard/logic/tools/science/numeral_bases.dart';
+import 'package:gc_wizard/logic/tools/science_and_technology/numeral_bases.dart';
 import 'package:gc_wizard/widgets/common/base/gcw_textfield.dart';
 import 'package:gc_wizard/widgets/common/gcw_numeralbase_spinner.dart';
 import 'package:gc_wizard/widgets/common/gcw_default_output.dart';
