+++ lib/widgets/tools/science_and_technology/primes/primes_integerfactorization.dart
@@ -1,22 +1,18 @@
 import 'package:flutter/material.dart';
-import 'package:gc_wizard/i18n/app_localizations.dart';
-import 'package:gc_wizard/logic/tools/crypto/rotator.dart';
-import 'package:gc_wizard/logic/tools/science/primes/primes.dart';
+import 'package:gc_wizard/logic/tools/science_and_technology/primes/primes.dart';
 import 'package:gc_wizard/widgets/common/base/gcw_output_text.dart';
-import 'package:gc_wizard/widgets/common/base/gcw_textfield.dart';
 import 'package:gc_wizard/widgets/common/gcw_integer_spinner.dart';
-import 'package:gc_wizard/widgets/common/gcw_default_output.dart';
 import 'package:gc_wizard/widgets/common/gcw_output.dart';
 
-class NearestPrime extends StatefulWidget {
+class IntegerFactorization extends StatefulWidget {
   @override
-  NearestPrimeState createState() => NearestPrimeState();
+  IntegerFactorizationState createState() => IntegerFactorizationState();
 }
 
-class NearestPrimeState extends State<NearestPrime> {
-  String _output = '2';
+class IntegerFactorizationState extends State<IntegerFactorization> {
+  String _output = '1';
 
-  var _currentPrimes = [2];
+  List<BigInt> _currentFactors = [BigInt.one];
 
   @override
   void initState() {
@@ -29,9 +25,10 @@ class NearestPrimeState extends State<NearestPrime> {
       children: <Widget>[
         GCWIntegerSpinner(
           value: 1,
+          min: 1,
           onChanged: (value) {
             setState(() {
-              _currentPrimes = getNearestPrime(value);
+              _currentFactors = integerFactorization(value);
             });
           },
         ),
@@ -41,12 +38,9 @@ class NearestPrimeState extends State<NearestPrime> {
   }
 
   Widget _buildOutput() {
-    if (_currentPrimes == null)
-      _currentPrimes = [];
-
     return GCWOutput(
       child: Column(
-        children: _currentPrimes.map((prime) => GCWOutputText(
+        children: _currentFactors.map((prime) => GCWOutputText(
           text: prime.toString()
         )).toList(),
       )
