+++ lib/widgets/tools/science_and_technology/primes/primes_isprime.dart
@@ -1,17 +1,15 @@
 import 'package:flutter/material.dart';
 import 'package:gc_wizard/i18n/app_localizations.dart';
-import 'package:gc_wizard/logic/tools/crypto/rotator.dart';
-import 'package:gc_wizard/logic/tools/science/primes/primes.dart';
-import 'package:gc_wizard/widgets/common/base/gcw_textfield.dart';
-import 'package:gc_wizard/widgets/common/gcw_integer_spinner.dart';
+import 'package:gc_wizard/logic/tools/science_and_technology/primes/primes.dart';
 import 'package:gc_wizard/widgets/common/gcw_default_output.dart';
+import 'package:gc_wizard/widgets/common/gcw_integer_spinner.dart';
 
-class PrimeIndex extends StatefulWidget {
+class IsPrime extends StatefulWidget {
   @override
-  PrimeIndexState createState() => PrimeIndexState();
+  IsPrimeState createState() => IsPrimeState();
 }
 
-class PrimeIndexState extends State<PrimeIndex> {
+class IsPrimeState extends State<IsPrime> {
   String _output = '';
 
   @override
@@ -25,12 +23,9 @@ class PrimeIndexState extends State<PrimeIndex> {
       children: <Widget>[
         GCWIntegerSpinner(
           value: 1,
-          min: 1,
-          max: 1000003,
           onChanged: (value) {
             setState(() {
-              var index = getPrimeIndex(value);
-              _output = index >= 1 ? index.toString() : i18n(context, 'primes_noprime');
+              _output = isPrime(BigInt.from(value)) ? i18n(context, 'primes_isprime_isprime') : i18n(context, 'primes_noprime');
             });
           },
         ),
