+++ lib/widgets/tools/science_and_technology/primes/primes_nearestprime.dart
@@ -1,22 +1,18 @@
 import 'package:flutter/material.dart';
-import 'package:gc_wizard/i18n/app_localizations.dart';
-import 'package:gc_wizard/logic/tools/crypto/rotator.dart';
-import 'package:gc_wizard/logic/tools/science/primes/primes.dart';
+import 'package:gc_wizard/logic/tools/science_and_technology/primes/primes.dart';
 import 'package:gc_wizard/widgets/common/base/gcw_output_text.dart';
-import 'package:gc_wizard/widgets/common/base/gcw_textfield.dart';
 import 'package:gc_wizard/widgets/common/gcw_integer_spinner.dart';
-import 'package:gc_wizard/widgets/common/gcw_default_output.dart';
 import 'package:gc_wizard/widgets/common/gcw_output.dart';
 
-class IntegerFactorization extends StatefulWidget {
+class NearestPrime extends StatefulWidget {
   @override
-  IntegerFactorizationState createState() => IntegerFactorizationState();
+  NearestPrimeState createState() => NearestPrimeState();
 }
 
-class IntegerFactorizationState extends State<IntegerFactorization> {
-  String _output = '1';
+class NearestPrimeState extends State<NearestPrime> {
+  String _output = '2';
 
-  List<BigInt> _currentFactors = [BigInt.one];
+  var _currentPrimes = [2];
 
   @override
   void initState() {
@@ -29,10 +25,9 @@ class IntegerFactorizationState extends State<IntegerFactorization> {
       children: <Widget>[
         GCWIntegerSpinner(
           value: 1,
-          min: 1,
           onChanged: (value) {
             setState(() {
-              _currentFactors = integerFactorization(value);
+              _currentPrimes = getNearestPrime(value);
             });
           },
         ),
@@ -42,9 +37,12 @@ class IntegerFactorizationState extends State<IntegerFactorization> {
   }
 
   Widget _buildOutput() {
+    if (_currentPrimes == null)
+      _currentPrimes = [];
+
     return GCWOutput(
       child: Column(
-        children: _currentFactors.map((prime) => GCWOutputText(
+        children: _currentPrimes.map((prime) => GCWOutputText(
           text: prime.toString()
         )).toList(),
       )
