+++ test/logic/tools/science_and_technology/primes/primes.dart
@@ -1,5 +1,5 @@
 import "package:flutter_test/flutter_test.dart";
-import 'package:gc_wizard/logic/tools/science/primes/primes.dart';
+import 'package:gc_wizard/logic/tools/science_and_technology/primes/primes.dart';
 
 void main() {
   group("Primes.getNthPrime:", () {
